renamed sedov_generator to sedov_constants
